Added CPU waste

+++ src/main/java/com/ebberod/trader_annotated/MarketLink.java
@@ -45,8 +45,8 @@ public class MarketLink {
      */
     @Timed(absolute = true)
     private String waitReply(String token) throws InterruptedException {
-        Thread.sleep((long) Math.abs(r.nextGaussian()) * 10 +
-                50 * ( 1 + (long) Math.sin(2 * Math.PI * System.currentTimeMillis() / 86400000)));
+        Thread.sleep((long) (Math.abs(r.nextGaussian()) * 10 +
+                (50 * ( 1 + Math.sin(2 * Math.PI * System.currentTimeMillis() / 86400000)))));
         return r.nextInt(10) > 8 ? "FAILURE" : "OK";
     }
 }
